Adding in media loading at game start.

+++ app/src/main/java/edu/uoregon/casls/aris_android/data_objects/Dialog.java
@@ -6,12 +6,12 @@ import edu.uoregon.casls.aris_android.Utilities.AppConfig;
  * Created by smorison on 8/19/15.
  */
 public class Dialog {
-	public int    dialog_id              = 0;
+	public long    dialog_id              = 0;
 	public String name                   = "Dialog";
 	public String desc                   = "";
-	public int    icon_media_id          = 0;
-	public int    intro_dialog_script_id = 0;
-	public int    back_button_enabled    = 1;
+	public long    icon_media_id          = 0;
+	public long    intro_dialog_script_id = 0;
+	public long    back_button_enabled    = 1;
 
 	public Dialog copy() {
 		Dialog c = new Dialog();
@@ -32,7 +32,7 @@ public class Dialog {
 		return "Dialog- Id:" + this.dialog_id + "\tName:" + this.name + "\t";
 	}
 
-	public int icon_media_id() {
+	public long icon_media_id() {
 		if (icon_media_id == 0) return Media.DEFAULT_DIALOG_ICON_MEDIA_ID;
 		return icon_media_id;
 	}
